fix: Fixes function of adhar card and pan card function for enin
