+++ pkg/mocai/entities/pan/generator.go
@@ -1,22 +1,9 @@
-package helper
+package pan
 
 import (
 	"math/rand"
-	"strconv"
 	"time"
 )
-
-// Generate a random Aadhaar number
-func GenerateAadhaarNumber() string {
-	src := rand.NewSource(time.Now().UnixNano())
-	rng := rand.New(src)
-	aadhaar := ""
-
-	for i := 0; i < 12; i++ {
-		aadhaar += strconv.Itoa(rng.Intn(10))
-	}
-	return aadhaar
-}
 
 // Function to generate a random PAN number
 func GenerateRandomPAN() string {
